Fix for bugs 100, 101, 102.
Fix for 100/101:
If lsof or strace can not be invoked or either returns a non-zero exit code, then ioprofile will exit and  report the error.

Fix for 102:
ioprofile now sleeps for one second after invoking strace and then checks for the existance of the strace process.  The strace should be running in the background if everything is working as expected.  If the strace is not running, then the contents of ${OPT_k} are displayed as these now include any error messages since I added 2>&1 to the lsof and strace invocations.